New Updated Level 1
I included the updated level 1 in a new scene "Level 1 - 2.0". I tagged all of the objects, placed the box colliders, and placed the scripts as best as I could. I also made some changes to my GuardMovementScript that caused errors when the guard waypoints were not set. I still need to set the guard waypoints and there are several errors that need to be fixed that I will discuss at the meeting tonight, but they should only affect the new scene I created.